*** This is a very bold move: possible breakages all over ***
At the very least, developers shall do a cold start:
    cd cctbx_build
    /usr/bin/python ../cctbx_project/libtbx/configure.py [options] phenix ...

*** In more details ***

Relocatable a cctbx becomes the default. The command line option '--relocatable' for configure therefore disappears. As a result, on Windows, the binary dispatchers are also a thing of the past.

This is achieved by replacing each and every string representing a path in the configuration layer by an instance of a new class relocatable_path. All those instances share the same ``root'' (the directory one level above cctbx_build) with libtbx.env, which is obtained from the OS environment variable LIBTBX_BUILD whether a warm or a cold start is performed.

However, I have tried to minimise disruption of anything but env_config.py itself: calls such as libtbx.env.under_build(...) still return a string as opposed to an instance of relocatable_path. There are only a few instances of command-line tools and SConscript which are broken because they directly refer to attributes of libtbx.env that now hold an instance of relocatable_path instead of a string.

*** Commits are to follow to sort out those issues. ***

Fixes are pretty mechanical: p being formerly a string representing a path and now an instance of relocatable_path,
    p = abs(p)
should cure the problem.

git-svn-id: http://svn.code.sf.net/p/cctbx/code/trunk@14235 3c1a6239-ac2c-0410-bc49-92500e444112