Silence harmless Clang warning.
Note: any code generated by fable could potentially trigger that warning.
The question is: how to read a && b || c?
Or in the case at hand a && b || c && d? Answer: && has a higher precedence
than || but a consensus has emerged in the C/C++ programming community that
too many people don't know that. So such expression trigger a warning and
one is supposed to explicitly spell out the precedence with parentheses.
In case of code generated by fable, better silence the warning.

git-svn-id: http://svn.code.sf.net/p/cctbx/code/trunk@15952 3c1a6239-ac2c-0410-bc49-92500e444112