Replace relative paths by absolute ones in error messages.
Motivation: the C++ source files that are generated within cctbx_build
are processed by SCons with their paths relative to that build directory
whereas the C++ source files inside cctbx_project are handled by SCons
with their absolute path. Unfortunately, Xcode does not recognize compiler
errors that feature relative a path. As a result, the click-to-jump to the
error location does not work for C++ source files generated inside cctbx_build
which make it more difficult to debug the code generation. Hence this script.

Note: this script is now entirely written in Python, using the
subprocess module to create a parent/child pair: the child process runs SCons
while the parent process filters the output of the child.

git-svn-id: http://svn.code.sf.net/p/cctbx/code/trunk@15955 3c1a6239-ac2c-0410-bc49-92500e444112