libtbx_refresh.py should not use libtbx.env
Each libtbx_refresh.py is like the body of a method of the class
libtbx.env_config.module as within each libtbx_refresh.py the
variable "self" is available at file scope that refers to the
instance of libtbx.env_config.module that is currently running
in libtbx/env_config.py.

Thus self.env shall be used instead of libtbx.env: forgetting to
do so may result in the wrong environment being updated.

git-svn-id: http://svn.code.sf.net/p/cctbx/code/trunk@16123 3c1a6239-ac2c-0410-bc49-92500e444112