Arithmetic operator simplification and add += and -=
Previously the code tried to factor common logic with a kind of
template metaprogramming. This reverts to a more pedestrian style of
programming with a bit of code duplication which is still far better
than the previous complexity.

Remove many of the friend and all of the static member functions
dealing with linear algebra operations and replace them with
non-static member functions. Keep the free functions`quadratic_form`
and `weighted_dot` for compatibility.

git-svn-id: http://svn.code.sf.net/p/cctbx/code/trunk@18223 3c1a6239-ac2c-0410-bc49-92500e444112