First pass at using XPP gain maps.
Metrology isn't used correctly yet.  More to come.

1) Modifies convert_gain_map to read an XPP ASCII gain map and write
it to a pickle file.  This file is displayable by cctbx.image_viewer.

2) Adds gain_map_level to the pyana module common_mode.  If present,
instead of using the supplied gain map directly, the gain level is
applied instead in a binary fashion, based on whether the pixels in
the gain map are zeros or ones.



git-svn-id: http://svn.code.sf.net/p/cctbx/code/trunk@18817 3c1a6239-ac2c-0410-bc49-92500e444112