Putative Format class for R-AXIS detector at SPring-8 BL26B1, which requires a reverse axis goniometer.
Warning: currently the understand method looks only for the string
"R-AXIS4" at the start of the file. This may well be true (seems 
likely anyway) for images written by other detectors that do not
need reversed goniometers, but we do not have example images from
such detectors yet.

It is unclear at this stage what could be used as a more specific
fingerprint for this instrument.



git-svn-id: http://svn.code.sf.net/p/cctbx/code/trunk@21457 3c1a6239-ac2c-0410-bc49-92500e444112