Keep original list order
This is an obscure one: reversing the list here causes the first two paths in
the environment variable PYTHONPATH to oscillate when running 'make reconf'
repeatedly. This is not a bad thing per se, but causes SCons to believe that
the dispatcher has changed. This in turn may lead to unnecessary build steps
to be run. At the moment there are no SCons nodes dependent on dispatchers,
at least I could not find any, so right now this should not change a thing.



git-svn-id: http://svn.code.sf.net/p/cctbx/code/trunk@22853 3c1a6239-ac2c-0410-bc49-92500e444112